requirements.txt
feedparser
discord.py
python-dotenv